backend/pyproject.toml
[project]
name = "urban-tree-observatory"
version = "0.1.0"
description = "Urban Tree Observatory"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-allauth[mfa,socialaccount]>=65.5.0",
    "django-cors-headers>=4.7.0",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-leaflet>=0.31.0",
    "django-storages[s3]>=1.14.5",
    "djangorestframework>=3.15.2",
    "djangorestframework-gis>=1.0",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "psycopg>=3.2.6",
    "sentry-sdk>=2.24.0",
    "tqdm>=4.67.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
    "coverage[toml]>=7.8.0",
    "django-debug-toolbar>=5.1.0",
    "django-extensions>=3.2.3",
    "factory-boy>=3.3.3",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-cov>=6.1.1",
    "pytest-django>=4.10.0",
    "ruff>=0.11.2",
]

[tool.pytest.ini_options]
minversion = "6.0"
DJANGO_SETTINGS_MODULE = "config.settings.dev"
addopts = "--cov=. --cov-report=term --reuse-db --import-mode=importlib"
python_files = "tests.py test_*.py *_tests.py"
python_classes = "Test* *Tests"
python_functions = "test_*"
testpaths = ["apps"]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["."]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "*/admin.py",
    "config/*",
    "manage.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "def __str__",
    "raise NotImplementedError",
    "if self.debug:",
    "if settings.DEBUG",
    "raise ImportError",
    "if 0:",
    "if __name__ == .__main__.:",
    "pass",
    "raise NotImplementedError"
]

[tool.coverage.html]
directory = "htmlcov"

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E203",  # whitespace before ':'
    "E501",  # line too long
    "B904",  # Within an `except` clause, raise exceptions with `raise ... from err` or `raise ... from None` to distinguish them from errors in exception handling
    "B017",  # Do not assert blind exception: `Exception`
]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
